requirement.txt
requests
beautifulsoup4
lxml
feedparser
google-generativeai
python-dotenv